Refactor + cleanup
